deployments-events-handler changes
