+++ core/jazz_deployments-event-handler/components/utils.js
@@ -20,14 +20,17 @@
     @author:
     @version: 1.0
 **/
-
+const _ = require("lodash");
+const logger = require("../components/logger.js");
 function checkForInterestedEvents(encodedPayload, sequenceNumber, configData) {
 	return new Promise((resolve, reject) => {
-		var kinesisPayload = JSON.parse(new Buffer(encodedPayload, 'base64').toString('ascii'));
+
+    var kinesisPayload = JSON.parse(new Buffer(encodedPayload, 'base64').toString('ascii'));
 		if (kinesisPayload.Item.EVENT_TYPE && kinesisPayload.Item.EVENT_TYPE.S) {
 			if (_.includes(configData.EVENTS.event_type, kinesisPayload.Item.EVENT_TYPE.S) &&
 				_.includes(configData.EVENTS.event_name, kinesisPayload.Item.EVENT_NAME.S)) {
-				logger.info("found " + kinesisPayload.Item.EVENT_TYPE.S + " event with sequence number: " + sequenceNumber);
+        logger.info("found " + kinesisPayload.Item.EVENT_TYPE.S + " event with sequence number: " + sequenceNumber);
+
 				return resolve({
 					"interested_event": true,
 					"payload": kinesisPayload.Item
@@ -43,8 +46,86 @@ function checkForInterestedEvents(encodedPayload, sequenceNumber, configData) {
 	});
 };
 
+function getDeploymentPayload(svcContext) {
+	var deploymentPayload = {};
+
+	if (svcContext.domain) {
+		deploymentPayload.domain = svcContext.domain;
+	}
+	if (svcContext.environment_logical_id) {
+		deploymentPayload.environment_logical_id = svcContext.environment_logical_id;
+	}
+	if (svcContext.provider_build_id) {
+		deploymentPayload.provider_build_id = svcContext.provider_build_id;
+	}
+	if (svcContext.provider_build_url) {
+		deploymentPayload.provider_build_url = svcContext.provider_build_url;
+	}
+	if (svcContext.scm_commit_hash) {
+		deploymentPayload.scm_commit_hash = svcContext.scm_commit_hash;
+	}
+	if (svcContext.scm_url) {
+		deploymentPayload.scm_url = svcContext.scm_url;
+	}
+	if (svcContext.scm_branch) {
+		deploymentPayload.scm_branch = svcContext.scm_branch;
+	}
+	if (svcContext.request_id) {
+		deploymentPayload.request_id = svcContext.request_id;
+	}
+	if (svcContext.status) {
+		deploymentPayload.status = svcContext.status;
+	}
+
+	return deploymentPayload;
+};
+
+function getSvcPayload(method, payload, apiEndpoint, authToken) {
+	var svcPayload = {
+		headers: {
+			'content-type': "application/json",
+			'authorization': authToken
+		},
+		rejectUnauthorized: false
+	}
+
+	svcPayload.uri = apiEndpoint;
+	svcPayload.method = method;
+	if (payload) {
+		svcPayload.json = payload;
+	}
+	logger.info("Deployment API payload :" + JSON.stringify(svcPayload));
+	return svcPayload;
+};
+function handleError(errorType, message) {
+	var error = {};
+	error.failure_code = errorType;
+	error.failure_message = message;
+	return error;
+};
+
+function getTokenRequest(configData) {
+	return {
+		uri: configData.BASE_API_URL + configData.TOKEN_URL,
+		method: 'post',
+		json: {
+			"username": configData.SERVICE_USER,
+			"password": configData.TOKEN_CREDS
+		},
+		rejectUnauthorized: false,
+		transform: (body, response, resolveWithFullResponse) => {
+			return response;
+		}
+	};
+}
+
+
 const exporatble  = {
-  checkForInterestedEvents
+  checkForInterestedEvents,
+  getDeploymentPayload,
+  getSvcPayload,
+  handleError,
+  getTokenRequest
 };
 module.exports =  exporatble;
