+++ core/jazz_deployments-event-handler/index.js
@@ -38,7 +38,7 @@ var failedEvents = [];
 function handler(event, context, cb) {
 	var config = configModule.getConfig(event, context);
 
-	rp(exportable.getTokenRequest(config))
+	rp(utils.getTokenRequest(config))
 		.then(result => {
 			return exportable.getAuthResponse(result);
 		})
@@ -55,22 +55,8 @@ function handler(event, context, cb) {
 			logger.error("Error processing events: " + JSON.stringify(err));
 			return cb(null, records);
 		});
-};
-
-function getTokenRequest(configData) {
-	return {
-		uri: configData.BASE_API_URL + configData.TOKEN_URL,
-		method: 'post',
-		json: {
-			"username": configData.SERVICE_USER,
-			"password": configData.TOKEN_CREDS
-		},
-		rejectUnauthorized: false,
-		transform: (body, response, resolveWithFullResponse) => {
-			return response;
-		}
-	};
 }
+
 
 function getAuthResponse(result) {
 	return new Promise((resolve, reject) => {
@@ -81,7 +67,7 @@ function getAuthResponse(result) {
 			return reject(errorHandler.throwInternalServerError("Invalid token response from API"));
 		}
 	});
-};
+}
 
 function processEventRecords(event, configData, authToken) {
 	return new Promise((resolve, reject) => {
@@ -98,7 +84,7 @@ function processEventRecords(event, configData, authToken) {
 				return reject(error);
 			});
 	});
-};
+}
 
 function processEventRecord(record, configData, authToken) {
 	return new Promise((resolve, reject) => {
@@ -126,7 +112,7 @@ function processEventRecord(record, configData, authToken) {
 				return reject(err);
 			});
 	});
-};
+}
 
 
 function processEvent(eventPayload, configData, authToken) {
@@ -151,95 +137,27 @@ function processEvent(eventPayload, configData, authToken) {
 			}
 		} else {
 			logger.info("Service Context is not defined");
-			var err = exportable.handleError(failureCodes.PR_ERROR_4.code, "Service context is not defined");
+			var err = utils.handleError(failureCodes.PR_ERROR_4.code, "Service context is not defined");
 			return reject(err);
 		}
 	});
-};
+}
 
-function handleError(errorType, message) {
-	var error = {};
-	error.failure_code = errorType;
-	error.failure_message = message;
-	return error;
-};
 
-function getDeploymentPayload(svcContext) {
-	var deploymentPayload = {};
 
-	if (svcContext.domain) {
-		deploymentPayload.domain = svcContext.domain;
-	}
-	if (svcContext.environment_logical_id) {
-		deploymentPayload.environment_logical_id = svcContext.environment_logical_id;
-	}
-	if (svcContext.provider_build_id) {
-		deploymentPayload.provider_build_id = svcContext.provider_build_id;
-	}
-	if (svcContext.provider_build_url) {
-		deploymentPayload.provider_build_url = svcContext.provider_build_url;
-	}
-	if (svcContext.scm_commit_hash) {
-		deploymentPayload.scm_commit_hash = svcContext.scm_commit_hash;
-	}
-	if (svcContext.scm_url) {
-		deploymentPayload.scm_url = svcContext.scm_url;
-	}
-	if (svcContext.scm_branch) {
-		deploymentPayload.scm_branch = svcContext.scm_branch;
-	}
-	if (svcContext.request_id) {
-		deploymentPayload.request_id = svcContext.request_id;
-	}
-	if (svcContext.status) {
-		deploymentPayload.status = svcContext.status;
-	}
 
-	return deploymentPayload;
-};
 
-function getSvcPayload(method, payload, apiEndpoint, authToken) {
-	var svcPayload = {
-		headers: {
-			'content-type': "application/json",
-			'authorization': authToken
-		},
-		rejectUnauthorized: false
-	}
-
-	svcPayload.uri = apiEndpoint;
-	svcPayload.method = method;
-	if (payload) {
-		svcPayload.json = payload;
-	}
-	logger.info("Deployment API payload :" + JSON.stringify(svcPayload));
-	return svcPayload;
-};
-
-function processRequest(svcPayload) {
-	return new Promise((resolve, reject) => {
-		request(svcPayload, function (error, response, body) {
-			if (response.statusCode === 200 && body) {
-				return resolve(body);
-			} else {
-				logger.error("Error processing request: " + JSON.stringify(response));
-				var error = exportable.handleError(failureCodes.PR_ERROR_3.code, response.body.message);
-				return reject(error);
-			}
-		});
-	});
-};
 
 function processCreateEvent(eventPayload, configData, authToken) {
 	return new Promise((resolve, reject) => {
 		var svcContext = JSON.parse(eventPayload.SERVICE_CONTEXT.S);
 		logger.info("svcContext: " + JSON.stringify(svcContext));
 
-		var deploymentPayload = exportable.getDeploymentPayload(svcContext);
+		var deploymentPayload = utils.getDeploymentPayload(svcContext);
 		deploymentPayload.service_id = eventPayload.SERVICE_ID.S
 		deploymentPayload.service = eventPayload.SERVICE_NAME.S
 		var apiEndpoint = configData.BASE_API_URL + configData.DEPLOYMENT_API_RESOURCE;
-		var svcPayload = exportable.getSvcPayload("POST", deploymentPayload, apiEndpoint, authToken);
+		var svcPayload = utils.getSvcPayload("POST", deploymentPayload, apiEndpoint, authToken);
 
 		exportable.processRequest(svcPayload)
 			.then(result => { return resolve(result); })
@@ -255,7 +173,7 @@ function processUpdateEvent(eventPayload, configData, authToken) {
 		var svcContext = JSON.parse(eventPayload.SERVICE_CONTEXT.S);
 		logger.info("svcContext: " + JSON.stringify(svcContext));
 
-		var deploymentPayload = exportable.getDeploymentPayload(svcContext);
+		var deploymentPayload = utils.getDeploymentPayload(svcContext);
 		deploymentPayload.service_id = eventPayload.SERVICE_ID.S
 		deploymentPayload.service = eventPayload.SERVICE_NAME.S
 
@@ -276,7 +194,7 @@ function getDeployments(deploymentPayload, configData, authToken) {
 			var service_name = deploymentPayload.service;
 			var domain = deploymentPayload.domain;
 			var apiEndpoint = configData.BASE_API_URL + configData.DEPLOYMENT_API_RESOURCE + "?service=" + service_name + "&domain=" + domain + "&environment=" + env_id;;
-			var svcPayload = exportable.getSvcPayload("GET", null, apiEndpoint, authToken);
+			var svcPayload = utils.getSvcPayload("GET", null, apiEndpoint, authToken);
 			exportable.processRequest(svcPayload)
 				.then(result => { return resolve(result); })
 				.catch(err => {
@@ -285,11 +203,11 @@ function getDeployments(deploymentPayload, configData, authToken) {
 				});
 		} else {
 			logger.error("Environment logical id is not defined");
-			var err = handleError(failureCodes.PR_ERROR_4.code, "Environment logical id is not defined");
+			var err = utils.handleError(failureCodes.PR_ERROR_4.code, "Environment logical id is not defined");
 			return reject(err);
 		}
 	});
-};
+}
 
 function updateDeployments(res, deploymentPayload, configData, authToken) {
 	return new Promise((resolve, reject) => {
@@ -312,7 +230,7 @@ function updateDeployments(res, deploymentPayload, configData, authToken) {
 				});
 				logger.info("Update deployment request payload: " + JSON.stringify(deploymentPayload));
 				var apiEndpoint = configData.BASE_API_URL + configData.DEPLOYMENT_API_RESOURCE + "/" + deploymentData.deployment_id;
-				var svcPayload = getSvcPayload("PUT", deploymentPayload, apiEndpoint, authToken);
+				var svcPayload = utils.getSvcPayload("PUT", deploymentPayload, apiEndpoint, authToken);
 				exportable.processRequest(svcPayload)
 					.then(result => { return resolve(result); })
 					.catch(err => {
@@ -321,30 +239,29 @@ function updateDeployments(res, deploymentPayload, configData, authToken) {
 					});
 			} else {
 				logger.error("Deployment details not found!");
-				var err = handleError(failureCodes.PR_ERROR_4.code, "Deployment details not found!");
+				var err = utils.handleError(failureCodes.PR_ERROR_4.code, "Deployment details not found!");
 				return reject(err);
 			}
 		} else {
 			logger.error("Deployment details not found!");
-			var err = handleError(failureCodes.PR_ERROR_4.code, "Deployment details not found!");
+			var err = utils.handleError(failureCodes.PR_ERROR_4.code, "Deployment details not found!");
 			return reject(err);
 		}
 	});
-};
+}
 
 function getEventProcessStatus() {
 	return {
 		"processed_events": processedEvents.length,
 		"failed_events": failedEvents.length
 	};
-};
-
+}
 function handleProcessedEvents(id, payload) {
 	processedEvents.push({
 		"sequence_id": id,
 		"event": payload
 	});
-};
+}
 
 function handleFailedEvents(id, failure_message, payload, failure_code) {
 	failedEvents.push({
@@ -353,12 +270,22 @@ function handleFailedEvents(id, failure_message, payload, failure_code) {
 		"failure_code": failure_code,
 		"failure_message": failure_message
 	});
+}
+function processRequest(svcPayload) {
+	return new Promise((resolve, reject) => {
+		request(svcPayload, function (error, response, body) {
+			if (response.statusCode === 200 && body) {
+				return resolve(body);
+			} else {
+				logger.error("Error processing request: " + JSON.stringify(response));
+				var error = utils.handleError(failureCodes.PR_ERROR_3.code, response.body.message);
+				return reject(error);
+			}
+		});
+	});
 };
-
 const exportable = {
-	getTokenRequest,
 	getAuthResponse,
-	handleError,
 	handleProcessedEvents,
 	handleFailedEvents,
 	getEventProcessStatus,
@@ -366,8 +293,6 @@ const exportable = {
 	processEventRecords,
 	processEventRecord,
 	processEvent,
-	getDeploymentPayload,
-	getSvcPayload,
 	processRequest,
 	processCreateEvent,
 	processUpdateEvent,
